Improve for ServerSesssion
