+++ src/lib.rs
@@ -8,19 +8,19 @@ extern crate tokio_io;
 extern crate bytes;
 extern crate iovec;
 
-
 mod common;
-mod tokio_impl;
+pub mod client;
+pub mod server;
 
-use std::mem;
-use std::io::{ self, Write };
+use std::{ io, mem };
 use std::sync::Arc;
 use webpki::DNSNameRef;
 use rustls::{
-    Session, ClientSession, ServerSession,
+    ClientSession, ServerSession,
     ClientConfig, ServerConfig
 };
-use tokio_io::{ AsyncRead, AsyncWrite };
+use futures::{Async, Future, Poll};
+use tokio_io::{ AsyncRead, AsyncWrite, try_nb };
 use common::Stream;
 
 
@@ -74,15 +74,15 @@ impl TlsConnector {
         f(&mut session);
 
         Connect(if self.early_data {
-            MidHandshake::EarlyData(TlsStream {
+            client::MidHandshake::EarlyData(client::TlsStream {
                 session, io: stream,
-                state: TlsState::EarlyData,
+                state: client::TlsState::EarlyData,
                 early_data: (0, Vec::new())
             })
         } else {
-            MidHandshake::Handshaking(TlsStream {
+            client::MidHandshake::Handshaking(client::TlsStream {
                 session, io: stream,
-                state: TlsState::Stream,
+                state: client::TlsState::Stream,
                 early_data: (0, Vec::new())
             })
         })
@@ -106,10 +106,9 @@ impl TlsAcceptor {
         let mut session = ServerSession::new(&self.inner);
         f(&mut session);
 
-        Accept(MidHandshake::Handshaking(TlsStream {
+        Accept(server::MidHandshake::Handshaking(server::TlsStream {
             session, io: stream,
-            state: TlsState::Stream,
-            early_data: (0, Vec::new())
+            state: server::TlsState::Stream,
         }))
     }
 }
@@ -117,182 +116,28 @@ impl TlsAcceptor {
 
 /// Future returned from `ClientConfigExt::connect_async` which will resolve
 /// once the connection handshake has finished.
-pub struct Connect<IO>(MidHandshake<IO, ClientSession>);
+pub struct Connect<IO>(client::MidHandshake<IO>);
 
 /// Future returned from `ServerConfigExt::accept_async` which will resolve
 /// once the accept handshake has finished.
-pub struct Accept<IO>(MidHandshake<IO, ServerSession>);
-
-enum MidHandshake<IO, S> {
-    Handshaking(TlsStream<IO, S>),
-    EarlyData(TlsStream<IO, S>),
-    End
-}
+pub struct Accept<IO>(server::MidHandshake<IO>);
 
 
-/// A wrapper around an underlying raw stream which implements the TLS or SSL
-/// protocol.
-#[derive(Debug)]
-pub struct TlsStream<IO, S> {
-    io: IO,
-    session: S,
-    state: TlsState,
-    early_data: (usize, Vec<u8>)
-}
+impl<IO: AsyncRead + AsyncWrite> Future for Connect<IO> {
+    type Item = client::TlsStream<IO>;
+    type Error = io::Error;
 
-#[derive(Debug)]
-enum TlsState {
-    EarlyData,
-    Stream,
-    Eof,
-    Shutdown
-}
-
-impl<IO, S> TlsStream<IO, S> {
-    #[inline]
-    pub fn get_ref(&self) -> (&IO, &S) {
-        (&self.io, &self.session)
-    }
-
-    #[inline]
-    pub fn get_mut(&mut self) -> (&mut IO, &mut S) {
-        (&mut self.io, &mut self.session)
-    }
-
-    #[inline]
-    pub fn into_inner(self) -> (IO, S) {
-        (self.io, self.session)
+    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
+        self.0.poll()
     }
 }
 
-impl<IO> io::Read for TlsStream<IO, ClientSession>
-where IO: AsyncRead + AsyncWrite
-{
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        let mut stream = Stream::new(&mut self.io, &mut self.session);
+impl<IO: AsyncRead + AsyncWrite> Future for Accept<IO> {
+    type Item = server::TlsStream<IO>;
+    type Error = io::Error;
 
-        match self.state {
-            TlsState::EarlyData => {
-                let (pos, data) = &mut self.early_data;
-
-                // complete handshake
-                if stream.session.is_handshaking() {
-                    stream.complete_io()?;
-                }
-
-                // write early data (fallback)
-                if !stream.session.is_early_data_accepted() {
-                    while *pos < data.len() {
-                        let len = stream.write(&data[*pos..])?;
-                        *pos += len;
-                    }
-                }
-
-                // end
-                self.state = TlsState::Stream;
-                data.clear();
-                stream.read(buf)
-            },
-            TlsState::Stream => match stream.read(buf) {
-                Ok(0) => {
-                    self.state = TlsState::Eof;
-                    Ok(0)
-                },
-                Ok(n) => Ok(n),
-                Err(ref e) if e.kind() == io::ErrorKind::ConnectionAborted => {
-                    self.state = TlsState::Shutdown;
-                    stream.session.send_close_notify();
-                    Ok(0)
-                },
-                Err(e) => Err(e)
-            },
-            TlsState::Eof | TlsState::Shutdown => Ok(0),
-        }
-    }
-}
-
-impl<IO> io::Read for TlsStream<IO, ServerSession>
-where IO: AsyncRead + AsyncWrite
-{
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        let mut stream = Stream::new(&mut self.io, &mut self.session);
-
-        match self.state {
-            TlsState::Stream => match stream.read(buf) {
-                Ok(0) => {
-                    self.state = TlsState::Eof;
-                    Ok(0)
-                },
-                Ok(n) => Ok(n),
-                Err(ref e) if e.kind() == io::ErrorKind::ConnectionAborted => {
-                    self.state = TlsState::Shutdown;
-                    stream.session.send_close_notify();
-                    Ok(0)
-                },
-                Err(e) => Err(e)
-            },
-            TlsState::Eof | TlsState::Shutdown => Ok(0),
-            TlsState::EarlyData => unreachable!()
-        }
-    }
-}
-
-impl<IO> io::Write for TlsStream<IO, ClientSession>
-where IO: AsyncRead + AsyncWrite
-{
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        let mut stream = Stream::new(&mut self.io, &mut self.session);
-
-        match self.state {
-            TlsState::EarlyData => {
-                let (pos, data) = &mut self.early_data;
-
-                // write early data
-                if let Some(mut early_data) = stream.session.early_data() {
-                    let len = early_data.write(buf)?;
-                    data.extend_from_slice(&buf[..len]);
-                    return Ok(len);
-                }
-
-                // complete handshake
-                if stream.session.is_handshaking() {
-                    stream.complete_io()?;
-                }
-
-                // write early data (fallback)
-                if !stream.session.is_early_data_accepted() {
-                    while *pos < data.len() {
-                        let len = stream.write(&data[*pos..])?;
-                        *pos += len;
-                    }
-                }
-
-                // end
-                self.state = TlsState::Stream;
-                data.clear();
-                stream.write(buf)
-            },
-            _ => stream.write(buf)
-        }
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Stream::new(&mut self.io, &mut self.session).flush()?;
-        self.io.flush()
-    }
-}
-
-impl<IO> io::Write for TlsStream<IO, ServerSession>
-where IO: AsyncRead + AsyncWrite
-{
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        let mut stream = Stream::new(&mut self.io, &mut self.session);
-        stream.write(buf)
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Stream::new(&mut self.io, &mut self.session).flush()?;
-        self.io.flush()
+    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
+        self.0.poll()
     }
 }
